requirements.txt
streamlit
numpy
Pillow
opencv-python-headless
scikit-image
psutil